src/control-center/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
Faker==18.12.0
